Refactor(screener): Separate orchestration from analysis logic
Moved the stage screening logic from `screener.py` to a new function `run_stage_screening` within `run_stage_analyzer.py`.

- `screener.py` is now a high-level orchestrator that calls the stage analyzer and then the base analyzer.
- `run_stage_analyzer.py` now contains both the original historical simulation logic and the new screening logic.
- This change improves modularity and removes code duplication, as requested.
- Added `pytz` to `requirements.txt` as it's a new dependency.
- Updated `.gitignore` to properly exclude generated output files and temporary test files.
