+++ .gitignore
@@ -134,7 +134,6 @@ stage_history_*/
 *.json
 
 # Screener output files
-stage*.csv
-*tradingview.txt
-*-screener_results.csv
-*-base.csv
+[0-9][0-9][0-9][0-9][0-9][0-9][0-9][0-9]-*.csv
+[0-9][0-9][0-9][0-9][0-9][0-9][0-9][0-9]-*.txt
+sampled_stocks.csv
