+++ sampled_stocks.csv
@@ -1,101 +1,101 @@
 Ticker,Exchange
-SMXT,NASDAQ
-YELP,NYSE
-KTOS,NASDAQ
-NTRS,NASDAQ
-FSHP,NASDAQ
-MJID,NASDAQ
-DKS,NYSE
-PRU,NYSE
-NAN,NYSE
-RFI,NYSE
-LGIH,NASDAQ
-GASS,NASDAQ
-CVGW,NASDAQ
-CHY,NASDAQ
-FIX,NYSE
-GILD,NASDAQ
-UNP,NYSE
-FER,NASDAQ
-MREO,NASDAQ
-BNGO,NASDAQ
-FCT,NYSE
-ETSY,NASDAQ
-ATHR,NASDAQ
-PRG,NYSE
-GRX,NYSE
-NEWT,NASDAQ
-HQL,NYSE
-FLYW,NASDAQ
-TTNP,NASDAQ
-VGI,NYSE
-TEAD,NASDAQ
-NBIX,NASDAQ
-ESLT,NASDAQ
-NEXN,NASDAQ
-NEU,NYSE
-XHG,NASDAQ
-TRGP,NYSE
-VRM,NASDAQ
-IMXI,NASDAQ
-APAM,NYSE
-NXGL,NASDAQ
-GDDY,NYSE
-HY,NYSE
-STKE,NASDAQ
-DCI,NYSE
-STFS,NASDAQ
-EHC,NYSE
-CDRO,NASDAQ
-NKTX,NASDAQ
-SSRM,NASDAQ
-AMCX,NASDAQ
-NVX,NASDAQ
-PSIX,NASDAQ
-SAIA,NASDAQ
-MIAX,NYSE
-FCF,NYSE
-RSF,NYSE
-OTF,NYSE
-CIFR,NASDAQ
-AUGO,NASDAQ
-TXT,NYSE
-ABBV,NYSE
-CX,NYSE
-MTVA,NASDAQ
-LTH,NYSE
-IPM,NASDAQ
-AIT,NYSE
-META,NASDAQ
-WYNN,NASDAQ
-FIBK,NASDAQ
+FHB,NASDAQ
+RDN,NYSE
+TTGT,NASDAQ
+BMRA,NASDAQ
+TEF,NYSE
+LAKE,NASDAQ
+EE,NYSE
+LOPE,NASDAQ
+ITGR,NYSE
+ISRG,NASDAQ
+ESGL,NASDAQ
+MQY,NYSE
+VERU,NASDAQ
+ETR,NYSE
+MVIS,NASDAQ
+ZTEK,NASDAQ
+DRTS,NASDAQ
+PLUG,NASDAQ
+BGMS,NASDAQ
+HOPE,NASDAQ
+FLNG,NYSE
+GAM,NYSE
+IOTR,NASDAQ
+SAP,NYSE
+CHDN,NASDAQ
+USIO,NASDAQ
+SRRK,NASDAQ
+BAK,NYSE
+NMIH,NASDAQ
 ELTX,NASDAQ
-FN,NYSE
-SMTI,NASDAQ
-MNMD,NASDAQ
-GFS,NASDAQ
-CNDT,NASDAQ
-AWI,NYSE
-BY,NYSE
+XNCR,NASDAQ
+XRAY,NASDAQ
+RYOJ,NASDAQ
+AIZN,NYSE
+ALOT,NASDAQ
+RLI,NYSE
+BFST,NASDAQ
+VICI,NYSE
+ALEC,NASDAQ
+ZTO,NYSE
+IVVD,NASDAQ
+DFDV,NASDAQ
+SLB,NYSE
+NREF,NYSE
+FTI,NYSE
+FG,NYSE
+WTS,NYSE
+EOT,NYSE
+CLLS,NASDAQ
+RR,NASDAQ
+GEHC,NASDAQ
+EHI,NYSE
+CWST,NASDAQ
+SNCY,NASDAQ
+GEV,NYSE
+GWRE,NYSE
+BPRN,NASDAQ
+SYF,NYSE
+ASRV,NASDAQ
+QNRX,NASDAQ
+NEXT,NASDAQ
+EYPT,NASDAQ
+NCT,NASDAQ
+ARWR,NASDAQ
+OXBR,NASDAQ
+MFIC,NASDAQ
+III,NASDAQ
+BLKB,NASDAQ
+HASI,NYSE
+OSCX,NASDAQ
+NIM,NYSE
+STEP,NASDAQ
+GSUN,NASDAQ
+AVBC,NYSE
+VUZI,NASDAQ
+RRBI,NASDAQ
+ENO,NYSE
+GRMN,NYSE
+XERS,NASDAQ
+RLJ,NYSE
+WVVI,NASDAQ
+APM,NASDAQ
+STGW,NASDAQ
+USFD,NYSE
+LPG,NYSE
+EPAM,NYSE
 HUDI,NASDAQ
-HPAI,NASDAQ
-GOOG,NASDAQ
-LX,NASDAQ
-BLUW,NASDAQ
-CAN,NASDAQ
-TRI,NASDAQ
-PLNT,NYSE
-CUBB,NYSE
-IHD,NYSE
-IDYA,NASDAQ
-ZWS,NYSE
+ALDX,NASDAQ
+OST,NASDAQ
+DXC,NYSE
 BAFN,NASDAQ
-VLO,NYSE
-GPK,NYSE
+CHEF,NASDAQ
+GLAD,NASDAQ
+TARS,NASDAQ
+EBS,NYSE
+CUZ,NYSE
 MGPI,NASDAQ
-GES,NYSE
-DKI,NASDAQ
-MBWM,NASDAQ
-ORA,NYSE
-CRML,NASDAQ
-RSKD,NYSE
+KFRC,NYSE
+CTRE,NYSE
+TTC,NYSE
